requirements.txt
Flask
python-dotenv
atproto
requests
textblob